dcms/pyproject.toml
[tool.poetry]
name = "dcms"
version = "0.1.0"
description = "Discord content management system"
authors = ["Charles de Freitas <me@chasbob.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.0"
authlib = "^0.14.3"
itsdangerous = "^1.1.0"
sqlalchemy = "^1.3.19"
psycopg2 = "^2.8.6"
httpx = "^0.15.4"
starlette-authlib = "^0.1.6"
"discord.py" = "^1.5.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
